requirements.txt
azure-ai-formrecognizer==3.2.0
azure-storage-blob==12.8.0
python-dotenv==1.0.0
pymongo==4.11.1